Revert master to 1.6.0
This reverts
 - commit d9890525acc0c14ee20b2523da4873551c6a91f2
 - commit e9a6d9acadf2a954df70dd5c84786088976f0479

+++ modules/matterhorn-series-service-api/src/main/java/org/opencastproject/series/api/SeriesQuery.java
@@ -26,9 +26,9 @@ import java.util.Date;
 public class SeriesQuery {
 
   /** Maximum number of results returned */
-  protected int count;
+  protected long count;
   /** Start page number */
-  protected int startPage;
+  protected long startPage;
   /** Free text search */
   protected String text;
   /** Series id search */
@@ -108,7 +108,7 @@ public class SeriesQuery {
    * @param count
    * @return
    */
-  public SeriesQuery setCount(int count) {
+  public SeriesQuery setCount(long count) {
     this.count = count;
     return this;
   }
@@ -119,7 +119,7 @@ public class SeriesQuery {
    * @param startPage
    * @return
    */
-  public SeriesQuery setStartPage(int startPage) {
+  public SeriesQuery setStartPage(long startPage) {
     this.startPage = startPage;
     return this;
   }
@@ -344,7 +344,7 @@ public class SeriesQuery {
    *
    * @return
    */
-  public int getCount() {
+  public long getCount() {
     return count;
   }
 
@@ -353,7 +353,7 @@ public class SeriesQuery {
    *
    * @return
    */
-  public int getStartPage() {
+  public long getStartPage() {
     return startPage;
   }
